// ProyectoAngular/primerAngular/src/app/models/profesor.ts
export class Profesor {
    id: string="";
   cedula: string="";
   correoelectronico: string="";
   telefono: string="";
   telefonocelular: string="";
   fechanacimiento: string="";
   sexo: string="";
   direccion: string="";
   nombre: string="";
   apellidopaterno: string="";
   apellidomaterno: string="";
   nacionalidad: string="";
   idCarreras: string="";
   usuario: string="";


}

/*export interface CursoResponse{
    code: number;
    message: string;
    data: Cursos[];
}*/